requirements.txt
flask
requests
python-dotenv
pandas
pyarrow
beautifulsoup4
scikit-learn
